Enforce unique order app data (#1)
CoW Protocol uses [app
data](https://docs.cow.fi/cow-protocol/reference/core/intents/app-data)
for features that are defined at a protocol level but not on-chain. An
example of such features is
[hooks](https://docs.cow.fi/cow-protocol/reference/core/intents/hooks).

Right now, if an order is valid for the Milkman contract, the app data
of that order can be changed with another arbitrary value and the order
is still valid.
The changes in this PR cause the app data to be unique for each order
and determined at order creation time.

#### Changes to the test

The test on the current main branch is failing. It's a `for` loop that
runs the test on a list of token and stops when any part fails.

To make sure that no regression are introduced, we first split the
combined test into multiple tests, one per token. We then see [which
tests fail
(3/10)](https://github.com/fedgiac/milkman/actions/runs/11214884698/job/31170897796)
for each token and compare the result with the [failing tests after
these changes
(2/10)](https://github.com/fedgiac/milkman/actions/runs/11215044485/job/31171406386).
The amount of failing tests is about the same so there should be no
regression in the code covered by the test. Note however that the tests
aren't deterministic so this technically isn't guaranteed.

The vast majority of changes in `Milkman.t.sol` are just indentation
changes.

---------

Co-authored-by: mfw78 <53399572+mfw78@users.noreply.github.com>